refactor(errors): move error response handler to common handler package
