+++ src/main/java/com/yan/virtuallibrary/Users/controller/UserBookController.java
@@ -6,7 +6,6 @@ import com.yan.virtuallibrary.Users.dto.UserBookResponseDTO;
 import com.yan.virtuallibrary.Users.dto.UserBookUpdateDTO;
 import com.yan.virtuallibrary.Users.service.UserBookService;
 import jakarta.validation.Valid;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
