+++ src/main/java/com/yan/virtuallibrary/Users/dto/UserUpdateDTO.java
@@ -1,8 +1,5 @@
 package com.yan.virtuallibrary.Users.dto;
 
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-
 public record UserUpdateDTO(
         String name,
         String username,
